requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.10
astroid==2.5.6
attrs==21.2.0
autopep8==1.5.6
colorama==0.4.4
Django==3.1.2
django-filter==2.4.0
django-mysql==3.11.1
djangorestframework==3.12.4
drf-spectacular==0.17.0
excelrd==2.0.3
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
inflection==0.5.1
isort==5.6.4
jsonschema==3.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql==0.0.3
mysqlclient==2.0.3
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pylint==2.8.2
pymssql==2.2.1
PyMySQL==1.0.2
PyPDF2==1.26.0
pyrsistent==0.17.3
python-bidi==0.4.2
pytz==2020.1
PyYAML==5.4.1
raw-zlib==0.1.15
reportlab==3.5.67
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
uritemplate==3.0.1
webencodings==0.5.1
wheel==0.30.0
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.5
xlwt==1.3.0